add option to use in-memory database
